% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eskinIndex.R
\name{eskinIndex}
\alias{eskinIndex}
\title{Eskin Index for categorical data}
\usage{
eskinIndex(dat2, key = NULL, simm = FALSE)
}
\arguments{
\item{key}{a character representing the unique key name for observations.  For example, this could contain the name of patients in a health dataset.}

\item{simm}{logical indicating whether a similarity should be returned}

\item{dat}{the data}

\item{weights}{a character variable representing the name of the variable associated with the survey weights}
}
\value{
A matrix of pairwise dissimilarities
}
\description{
This function calculates the dissimilarity index based on Eskin
}
\details{
This calculates the pairwise dissimilarities for a dataset using the goodall dissimilarity
index first proposed by Eskin.  Access through catDist function
}
\author{
Joshua Agterberg
}
