% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goodallIndex.R
\name{goodallIndex}
\alias{goodallIndex}
\title{Goodall Index (Versions 1,2,3, and 4) for categorical data}
\usage{
goodallIndex(dat, key = NULL, weights = NULL, simm = FALSE,
  type = 1, diag = FALSE)
}
\arguments{
\item{dat}{the data}

\item{key}{a character representing the unique key name for observations.
For example, this could contain the name of patients in a health dataset.}

\item{weights}{a character variable representing the name of the variable
associated with the survey weights}

\item{simm}{logical indicating whether a similarity should be returned}

\item{type}{a number in 1,2,3, or 4 indicating the the type of goodall index}
}
\value{
A matrix of pairwise dissimilarities
}
\description{
This function calculates the dissimilarity index based on Goodall
}
\details{
This calculates the pairwise dissimilarities for a dataset using the goodall dissimilarity
index first proposed by Goodall.
}
\author{
Joshua Agterberg
}
