% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inverseOccurenceIndex.R
\name{inverseOccurenceIndex}
\alias{inverseOccurenceIndex}
\title{Dissimilarity Indices With or Without Survey Weights}
\usage{
inverseOccurenceIndex(dat, key = NULL, weights, simm = FALSE)
}
\arguments{
\item{dat}{the data}

\item{key}{a character representing the unique key name for observations.
For example, this could contain the name of patients in a health dataset.}

\item{weights}{a character variable representing the name of the variable
associated with the survey weights}

\item{simm}{logical indicating whether a similarity should be returned}
}
\value{
A matrix of pairwise dissimilarities
}
\description{
This function calculates the dissimilarity index based on Goodall
}
\details{
This calculates the pairwise dissimilarities for a dataset using the inverse occurence dissimilarity
index
}
\author{
Joshua Agterberg
}
