% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occurenceIndex.R
\name{occurenceIndex}
\alias{occurenceIndex}
\title{Occurence (or inverse occurence) index for categorical data}
\usage{
occurenceIndex(dat, key = NULL, weights, simm = FALSE,
  inverse = TRUE)
}
\arguments{
\item{dat}{the data}

\item{key}{a character representing the unique key name for observations.
For example, this could contain the name of patients in a health dataset.}

\item{weights}{a character variable representing the name of the variable
associated with the survey weights}

\item{simm}{logical indicating whether a similarity should be returned}

\item{inverse}{parameter indicating whether to return the inverse occurence frequency or the occurence frequency}
}
\value{
A matrix of pairwise dissimilarities
}
\description{
This function calculates the dissimilarity index based on the (inverse) occurence index
}
\details{
This calculates the pairwise dissimilarities for a dataset using the (inverse) occurence dissimilarity
index
}
\author{
Joshua Agterberg
}
