\name{catdist-package}
\alias{catdist-package}
\alias{catdist}
\docType{package}
\title{
A general framework for implementing distances for
categorical variables
}
\description{
Distances for categorical variables computed based on the general framework described in van de Velden, Iodice D'Enza, Markos and Caviccia (2023).
}
\details{
\tabular{ll}{
Package: \tab catdist\cr
Type: \tab Package\cr
<<<<<<< HEAD
Version: \tab 0.2.0\cr
Date: \tab 2023-05-11\cr
=======
Version: \tab 0.1.1\cr
Date: \tab 2022-12-12\cr
>>>>>>> 72e0c5b6d717da613d07c4e0170eb887c0fe354b
License: \tab GPL-3\cr
}
}
\author{
Michel van de Velden [aut],
Alfonso Iodice D' Enza [aut],
Angelos Markos [aut, cre],
Carlo Cavicchia [aut]
}
\references{
van de Velden, M., Iodice D'Enza, A., Markos, A., & Cavicchia (2023). A general framework for implementing distances for categorical variables (in preparation)
}

\keyword{ package }
