% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catDist.R
\name{catDist}
\alias{catDist}
\title{Dissimilarity Indices for Clustering}
\usage{
catDist(dat, method = "goodall1", key = NULL, weights = NULL,
  simm = FALSE)
}
\arguments{
\item{dat}{A data.frame or data.frame coercible object}

\item{method}{the method to be used}

\item{key}{the key of the observations}

\item{weights}{an optional vector containing weights}

\item{simm}{should a similarity index be returned?}
}
\description{
Dissimilarity Indices for Clustering
}
